[PREM-4] Split create transaction page to internal and external transaction. Enhance UI.
