+++ src/app/modules/profile/receiver-edit-modal/ReceiverEditModalForm.tsx
@@ -8,6 +8,8 @@ import {ReceiversListLoading} from '@/app/modules/profile/loading/ReceiversListL
 import {ProfileService as profileService} from '../core/_requests';
 import {useAuth} from '@/app/modules/auth';
 import {ReceiverModalContext} from '@/app/modules/profile/components/Receivers';
+import useNotification from "@/app/modules/notifications/useNotification";
+import {AlertColor} from "@mui/material";
 
 type Props = {
   receiver: ReceiverDto;
@@ -32,7 +34,7 @@ const ReceiverEditModalForm: FC<Props> = ({receiver, isReceiverLoading}) => {
   // @ts-ignore
   const {openAddReceiverModal} = useContext(ReceiverModalContext);
   const {currentUser} = useAuth();
-
+  const {setNotification} = useNotification();
   const [receiverToInsert] = useState<ReceiverDto>(receiver);
 
   const cancel = (withRefresh?: boolean) => {
@@ -59,6 +61,11 @@ const ReceiverEditModalForm: FC<Props> = ({receiver, isReceiverLoading}) => {
         }
       } catch (ex) {
         console.error(ex);
+        console.log('error', ex);
+        const notificationType: AlertColor = "error";
+        const errorMessage: string = ex.response.data['Error: '] || "Something went wrong!";
+        console.log("errorMessage", errorMessage);
+        setNotification(true, errorMessage, notificationType, () => {});
       } finally {
         setSubmitting(true);
         cancel(true);
@@ -98,48 +105,8 @@ const ReceiverEditModalForm: FC<Props> = ({receiver, isReceiverLoading}) => {
                 style={{backgroundImage: `url('${userAvatarImg}')`}}
               ></div>
               {/* end::Preview existing avatar */}
-
-              {/* begin::Label */}
-              {/* <label
-              className='btn btn-icon btn-circle btn-active-color-primary w-25px h-25px bg-body shadow'
-              data-kt-image-input-action='change'
-              data-bs-toggle='tooltip'
-              title='Change avatar'
-            >
-              <i className='bi bi-pencil-fill fs-7'></i>
-
-              <input type='file' name='avatar' accept='.png, .jpg, .jpeg' />
-              <input type='hidden' name='avatar_remove' />
-            </label> */}
-              {/* end::Label */}
-
-              {/* begin::Cancel */}
-              {/* <span
-              className='btn btn-icon btn-circle btn-active-color-primary w-25px h-25px bg-body shadow'
-              data-kt-image-input-action='cancel'
-              data-bs-toggle='tooltip'
-              title='Cancel avatar'
-            >
-              <i className='bi bi-x fs-2'></i>
-            </span> */}
-              {/* end::Cancel */}
-
-              {/* begin::Remove */}
-              {/* <span
-              className='btn btn-icon btn-circle btn-active-color-primary w-25px h-25px bg-body shadow'
-              data-kt-image-input-action='remove'
-              data-bs-toggle='tooltip'
-              title='Remove avatar'
-            >
-              <i className='bi bi-x fs-2'></i>
-            </span> */}
-              {/* end::Remove */}
             </div>
             {/* end::Image input */}
-
-            {/* begin::Hint */}
-            {/* <div className='form-text'>Allowed file types: png, jpg, jpeg.</div> */}
-            {/* end::Hint */}
           </div>
           {/* end::Input group */}
