+++ src/app/modules/profile/receiver-list-modal/ReceiverListModal.tsx
@@ -10,16 +10,21 @@ import {useThemeMode} from "@_metronic/partials";
 type Props = {
   setIsShowModal: (isShowModal: boolean) => void;
   isShow: boolean;
+  isInternal: boolean;
   handleOnConfirmBtn: (selectedReceiver: ReceiverDto | null) => void;
 };
-const ReceiverListModal: FC<Props> = ({isShow, setIsShowModal, handleOnConfirmBtn}) => {
+const ReceiverListModal: FC<Props> = ({isShow, setIsShowModal, handleOnConfirmBtn, isInternal}) => {
   const {currentUser} = useAuth();
   const {mode} = useThemeMode();
   const [receivers, setReceivers] = React.useState<ReceiverDto[]>([]);
   const [selectedReceiver, setSelectedReceiver] = React.useState<ReceiverDto | null>(null);
   useEffect(() => {
     profileService.getAllReceiversByUserId(currentUser?.id).then((data: ReceiverDto[]) => {
-      setReceivers(data);
+      if (isInternal) {
+        setReceivers(data.filter((receiver) => receiver.bankName === 'Premierebank'));
+      } else {
+        setReceivers(data.filter((receiver) => receiver.bankName !== 'Premierebank'));
+      }
     }).catch((error) => {
       console.log(error);
     });
@@ -59,55 +64,63 @@ const ReceiverListModal: FC<Props> = ({isShow, setIsShowModal, handleOnConfirmBt
               </div>
               {/* begin::Modal body */}
               <div className='modal-body scroll-y mx-5 mx-xl-15 my-7'>
-                <table className='table table-row-gray-300 align-middle gs-0 gy-4'>
-                  <thead>
-                  <tr className='fw-bold text-muted'>
-                    <th className='w-10px'>
-                    </th>
-                    <th className='min-w-150px text-center'>Nickname</th>
-                    <th className='min-w-140px text-center'>Credit Card</th>
-                  </tr>
-                  </thead>
-                  {/* begin::Table body */}
-                  <tbody>
-                  {receivers.map((receiver: ReceiverDto) => (
-                      <tr key={receiver.id}>
-                        <td>
-                          <div
-                              className='form-check form-check-sm form-check-custom form-check-solid'>
-                            <input className='form-check-input widget-9-check'
-                                   type='checkbox'
-                                   name={'receiverCheckbox'}
-                                   checked={selectedReceiver?.id === receiver.id}
-                                   onChange={() => {
-                                      setSelectedReceiver(receiver);
-                                   }}
-                            />
-                          </div>
-                        </td>
-                        <td>
-                          <div className='d-flex flex-column ms-3 text-center'>
-                            <a href='#' className='text-gray-800 text-hover-primary fs-6 fw-bolder'>
-                              {receiver.nickname}
-                            </a>
-                          </div>
-                        </td>
-                        <td>
-                          <div className='d-flex align-items-center'>
-                            <Card className={'creditcard'}
-                                  bankName={receiver.bankName}
-                                  cardHolder={receiver.fullName}
-                                  cardNumber={receiver.cardNumber}
-                                  issuer='visa'
-                                  theme={mode}
-                            />
-                          </div>
-                        </td>
-                      </tr>
-                  ))}
-                  </tbody>
-                  {/* end::Table body */}
-                </table>
+                {
+                  receivers.length > 0 ? (
+                      <table className='table table-row-gray-300 align-middle gs-0 gy-4'>
+                        <thead>
+                        <tr className='fw-bold text-muted'>
+                          <th className='w-10px'>
+                          </th>
+                          <th className='min-w-150px text-center'>Nickname</th>
+                          <th className='min-w-140px text-center'>Credit Card</th>
+                        </tr>
+                        </thead>
+                        {/* begin::Table body */}
+                        <tbody>
+                        {receivers.map((receiver: ReceiverDto) => (
+                            <tr key={receiver.id}>
+                              <td>
+                                <div
+                                    className='form-check form-check-sm form-check-custom form-check-solid'>
+                                  <input className='form-check-input widget-9-check'
+                                         type='checkbox'
+                                         name={'receiverCheckbox'}
+                                         checked={selectedReceiver?.id === receiver.id}
+                                         onChange={() => {
+                                           setSelectedReceiver(receiver);
+                                         }}
+                                  />
+                                </div>
+                              </td>
+                              <td>
+                                <div className='d-flex flex-column ms-3 text-center'>
+                                  <a href='#' className='text-gray-800 text-hover-primary fs-6 fw-bolder'>
+                                    {receiver.nickname}
+                                  </a>
+                                </div>
+                              </td>
+                              <td>
+                                <div className='d-flex align-items-center'>
+                                  <Card className={'creditcard'}
+                                        bankName={receiver.bankName}
+                                        cardHolder={receiver.fullName}
+                                        cardNumber={receiver.cardNumber}
+                                        issuer='visa'
+                                        theme={mode}
+                                  />
+                                </div>
+                              </td>
+                            </tr>
+                        ))}
+                        </tbody>
+                        {/* end::Table body */}
+                      </table>
+                  ) : (
+                      <div className='text-center'>
+                        <h3 className='text-gray-600'>No receiver found</h3>
+                      </div>
+                  )
+                }
                 <button type={'button'}
                         className={'btn btn-primary float-end'}
                         disabled={selectedReceiver === null}
